fix(windows): simplify post-onboarding controls
Move the existing Let Omi take actions toggle from Privacy to General without changing its behavior. Make Goal the terminal onboarding step and open Chat immediately after selection or skip.

Failure-Class: none
