+++ desktop/windows/src/renderer/src/components/settings/tabs/PrivacyTab.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react'
-import { Activity, EyeOff, Monitor, ShieldCheck, Zap } from 'lucide-react'
-import { getPreferences, setPreferences } from '../../../lib/preferences'
+import { Activity, EyeOff, Monitor, ShieldCheck } from 'lucide-react'
 import { SettingRow } from '../SettingRow'
 import { Toggle } from '../Toggle'
 import type { UsageSettings } from '../../../../../shared/types'
@@ -23,18 +22,6 @@ export function PrivacyTab(): React.JSX.Element {
   }, [])
   const saveUsage = async (next: UsageSettings): Promise<void> => {
     setUsage(await window.omi.usageSetSettings(next))
-  }
-
-  // Desktop-automation opt-in. The toggle writes the same `automationConsentedAt`
-  // preference the onboarding Automation step sets; useChat gates its action
-  // planner on it. `OMI_AUTOMATION=0` is a hard build kill-switch (exposed as
-  // automationEnabled) — when off, the feature can't run, so the toggle is shown
-  // disabled regardless of consent.
-  const automationAvailable = window.omi.automationEnabled
-  const [autoConsent, setAutoConsent] = useState<boolean>(!!getPreferences().automationConsentedAt)
-  const toggleAutomation = (on: boolean): void => {
-    setAutoConsent(on)
-    setPreferences({ automationConsentedAt: on ? Date.now() : undefined })
   }
 
   // Bar/HUD screen-share privacy: exclude the top-edge bar from captures
@@ -69,27 +56,6 @@ export function PrivacyTab(): React.JSX.Element {
 
   return (
     <>
-      <SettingRow
-        icon={Zap}
-        dot={automationAvailable && autoConsent ? 'on' : 'off'}
-        title="Let Omi take actions"
-        subtitle={
-          !automationAvailable
-            ? 'Disabled in this build.'
-            : autoConsent
-              ? 'Omi can click and type in your apps when you ask — you approve each action first.'
-              : 'Turn on to let Omi act in your apps when you ask (you approve each action first).'
-        }
-        keywords="automation actions desktop control agent take action flaui approve"
-        control={
-          <Toggle
-            on={automationAvailable && autoConsent}
-            onChange={toggleAutomation}
-            disabled={!automationAvailable}
-            label="Let Omi take actions"
-          />
-        }
-      />
       <SettingRow
         icon={Activity}
         dot={usage?.enabled ? 'on' : 'off'}
